+++ database/migrations/2019_06_13_075235_create_projects_table.php
@@ -17,7 +17,7 @@ class CreateProjectsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('info');
-            $table->integer('tarif');
+            $table->bigInteger('tarif');
             $table->timestamps();
         });
     }
